layout customization
first bout of advanced customization, changing layouts of certain branch pages, indexes, and content pages.

most notably are the changes to each of the main indexes (world, posts, devlog, & media), making a custom layout for each to have a unique header image (current header images are placeholders to be changed).

also changed the default profile picture to the new WiR logo (another two-tone placeholder to be changed for a gradient logo soon)