reupped author-extra
hugo builds fine on my machine, idk why the gitbhub server is having trouble. the errors are due to hugo unable to find the author pfp when the author-extra partial is called. i've changed around stuff in the partial in an attempt to let it load something to avoid a nil error